Chooses Food Based on Path Distance
Now path sometimes returns a string Unassigned when it can't find a
path towards food.

Also removes Tate's dumb random joke in the move method.

Depricates getNearestFood function. I'd like to create a folder in the
repo for old code so we can reuse it incase some of it is useful.
This was the efforts of Stephen and Kyle.